User can now enter address. A marker is put on map, marker is focused, and address added to list of addresses. Click on an address in the list and the marker corresponding to that address is focused. I have my API token for Mapbox on the client with no real protection. I made an attempt to initialize necessary objects etc. server side and send it back to client but couldn't get the flow right. I do not think someone would take my api token and make the max requests, but I see this as a vulnerability and will be researching how to solve this specific problem.
Mapbox has some instructions for securely using the API token. I will be reading into this
and may refactor code to be more secure.
As this a small application with not many dependencies, it may be best to solve this
problem in this scope. That way I can take what I learned from here to the next project
I work on.

If anyone sees this and has suggestions please let me know!
Thanks in advance,
Phil